+++ gradle-plugins/react/brownfield/src/main/kotlin/com/callstack/react/brownfield/processors/VariantTaskProvider.kt
@@ -3,6 +3,7 @@ package com.callstack.react.brownfield.processors
 import com.android.build.gradle.internal.tasks.factory.dependsOn
 import com.callstack.react.brownfield.exceptions.TaskNotFound
 import com.callstack.react.brownfield.shared.ExplodeAarTask
+import com.callstack.react.brownfield.shared.MergeClassesTask
 import com.callstack.react.brownfield.utils.AndroidArchiveLibrary
 import com.callstack.react.brownfield.utils.DirectoryManager
 import com.callstack.react.brownfield.utils.Extension
@@ -83,7 +84,7 @@ class VariantTaskProvider(val project: Project) {
         preBuildTask.dependsOn(explodeAarTask)
 
         val bundledAssetsVariantName =
-            Utils.getBundledAssetsVariantName(
+            VariantHelper.getBundledAssetsVariantName(
                 variantName = variantName,
                 buildTypeName = buildTypeName,
                 isDebuggable = isVariantDebuggable,
@@ -95,10 +96,28 @@ class VariantTaskProvider(val project: Project) {
         preBuildTask.dependsOn("${appProject.path}:createBundle${capitalizedBundledAssetsVariantName}JsAndAssets")
 
         if (Utils.isExpoProject(project)) {
-            val updatesResourcesTaskName = Utils.getExpoUpdatesResourcesTaskName(variantName)
+            val updatesResourcesTaskName = VariantHelper.getExpoUpdatesResourcesTaskName(variantName)
             if (Utils.hasExpoUpdates(appProject, variantName)) {
                 preBuildTask.dependsOn("${appProject.path}:$updatesResourcesTaskName")
             }
+        }
+    }
+
+    fun mergeClasses(
+        aarLibraries: Collection<AndroidArchiveLibrary>,
+        explodeTasks: TaskProvider<ExplodeAarTask>,
+        variantName: String,
+    ): TaskProvider<MergeClassesTask> {
+        val capitalizedVariantName = variantName.capitalized()
+        val mergeClassesTaskName = "mergeClasses$capitalizedVariantName"
+
+        return project.tasks.register(mergeClassesTaskName, MergeClassesTask::class.java) {
+            it.dependsOn(explodeTasks)
+            it.dependsOn(VariantHelper.getKotlinCompileTask(project, capitalizedVariantName))
+            it.dependsOn(VariantHelper.getJavaCompileTask(project, capitalizedVariantName))
+            it.variantName.set(variantName)
+            it.inputClassesJars.from(aarLibraries.map { aarLibrary -> aarLibrary.getClassesJarFile() })
+            it.outputDir.set(DirectoryManager.getMergeClassDirectory(variantName))
         }
     }
 }